robotframework-parser/pyproject.toml
[tool.poetry]
name = "robot-parser"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.scripts]
robot-parser = "robot_parser.parser:main"

[tool.poetry.dependencies]
python = "^3.8"
robotframework = "^3.2.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.5.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
